Renames and strip down, at present not-likely to run
