+++ pom.xml
@@ -9,12 +9,12 @@
 
   <groupId>io.seqware</groupId>
   <!-- MAKE SURE YOU UPDATE THE workflow.properties FILE TOO -->
-  <artifactId>SangerPancancerCgpCnIndelSnvStr</artifactId>
+  <artifactId>CgpSomaticCore</artifactId>
   <!-- MAKE SURE YOU UPDATE THE workflow.properties FILE TOO -->
-  <version>1.0.8</version>
+  <version>0.0.0</version>
   <packaging>jar</packaging>
   <!-- MAKE SURE YOU UPDATE THE workflow.properties FILE TOO -->
-  <name>SangerPancancerCgpCnIndelSnvStr</name>
+  <name>CgpSomaticCore</name>
   <url>http://maven.apache.org</url>
 
 	<properties>
@@ -22,12 +22,10 @@
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<seqware-version>1.1.0</seqware-version>
 		<skipITs>true</skipITs>
-		<workflow-full-name>Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr_${workflow-version}_SeqWare_${seqware-version}</workflow-full-name>
-		<workflow-dir>${project.build.directory}/${workflow-full-name}/Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr/${workflow-version}</workflow-dir>
+		<workflow-full-name>Workflow_Bundle_CgpSomaticCore_${workflow-version}_SeqWare_${seqware-version}</workflow-full-name>
+		<workflow-dir>${project.build.directory}/${workflow-full-name}/Workflow_Bundle_CgpSomaticCore/${workflow-version}</workflow-dir>
 		<version.template.file>src/main/java/io/seqware/pancancer/Version.java.template</version.template.file>
 		<version.file>src/main/java/io/seqware/pancancer/Version.java</version.file>
-		<gt-download-upload-wrapper.version>2.0.10</gt-download-upload-wrapper.version>
-		<vcf-uploader.version>2.0.4</vcf-uploader.version>
 	</properties>
 
 	<profiles>
@@ -66,12 +64,6 @@
 
 	<dependencies>
 		<dependency>
-		    <groupId>vcf-uploader</groupId>
-		    <artifactId>vcf-uploader</artifactId>
-		    <version>${vcf-uploader.version}</version>
-		    <type>zip</type>
-		</dependency>
-		<dependency>
 			<groupId>io.seqware</groupId>
 			<artifactId>seqware-bin-linux-x86-64-jre-7.0.51</artifactId>
 			<version>1.0.0</version>
@@ -88,12 +80,6 @@
 			<artifactId>SeqWare-CGP-SomaticCore_opt</artifactId>
 			<version>1.3.1</version>
 			<type>tar.gz</type>
-		</dependency>
-		<dependency>
-			<groupId>gt-download-upload-wrapper</groupId>
-			<artifactId>gt-download-upload-wrapper</artifactId>
-			<version>${gt-download-upload-wrapper.version}</version>
-			<type>zip</type>
 		</dependency>
 	</dependencies>
