+++ scripts/sanger_startup.sh
@@ -7,6 +7,6 @@ echo "Downloading refdata..."
 DEST_DIR=/refdata/data
 
 perl /home/seqware/Seqware-CGP-SomaticCore/workflow/scripts/download_ref.pl  $DEST_DIR
-ln -s $DEST_DIR /home/seqware/Seqware-CGP-SomaticCore/target/Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr_1.0.8_SeqWare_1.1.0/Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr/1.0.8/data
+ln -s $DEST_DIR /home/seqware/Seqware-CGP-SomaticCore/target/Workflow_Bundle_CgpSomaticCore_1.0.8_SeqWare_1.1.0/Workflow_Bundle_CgpSomaticCore/1.0.8/data
 
 exec $@
