+++ workflow.properties
@@ -1,6 +1,6 @@
-workflow-name=SangerPancancerCgpCnIndelSnvStr
-workflow-short-name-version=svcp_1-0-8
-workflow-version=1.0.8
+workflow-name=CgpSomaticCore
+workflow-short-name-version=csc_0-0-0
+workflow-version=0.0.0
 workflow-description=Runs the WTSI CGP Copynumber, Indel, Snv and Structural variation core pipeline.
-workflow-url=https://s3.amazonaws.com/oicr.workflow.bundles/released-bundles/Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr_1.0.8_SeqWare_1.1.0.zip
-workflow-src-url=https://github.com/ICGC-TCGA-PanCancer/SeqWare-CGP-SomaticCore
+workflow-url=https://s3.amazonaws.com/oicr.workflow.bundles/released-bundles/Workflow_Bundle_CgpSomaticCore_0.0.0_SeqWare_1.1.0.zip
+workflow-src-url=https://github.com/ICGC-TCGA-PanCancer/CGP-Somatic-Docker
